refactor: qc-core user 도메인 패키지 정리
