+++ qc-application/src/main/java/com/eager/questioncloud/application/api/authentication/AuthenticationProcessor.java
@@ -2,9 +2,9 @@ package com.eager.questioncloud.application.api.authentication;
 
 import com.eager.questioncloud.core.domain.social.SocialAPIManager;
 import com.eager.questioncloud.core.domain.social.SocialPlatform;
-import com.eager.questioncloud.core.domain.user.AccountType;
-import com.eager.questioncloud.core.domain.user.User;
-import com.eager.questioncloud.core.domain.user.UserRepository;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
+import com.eager.questioncloud.core.domain.user.model.User;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
