+++ qc-application/src/main/java/com/eager/questioncloud/application/api/authentication/AuthenticationService.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.authentication;
 
-import com.eager.questioncloud.core.domain.user.AccountType;
-import com.eager.questioncloud.core.domain.user.User;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
