+++ qc-application/src/main/java/com/eager/questioncloud/application/api/creator/CreatorRegister.java
@@ -3,8 +3,8 @@ package com.eager.questioncloud.application.api.creator;
 import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
 import com.eager.questioncloud.core.domain.creator.model.Creator;
 import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
-import com.eager.questioncloud.core.domain.user.User;
-import com.eager.questioncloud.core.domain.user.UserRepository;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
+import com.eager.questioncloud.core.domain.user.model.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
