+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/register/RegisterUserControllerRequest.java
@@ -3,8 +3,8 @@ package com.eager.questioncloud.application.api.user.register;
 import com.eager.questioncloud.application.validator.EmailValidator;
 import com.eager.questioncloud.application.validator.PasswordValidator;
 import com.eager.questioncloud.application.validator.Validatable;
-import com.eager.questioncloud.core.domain.user.AccountType;
-import com.eager.questioncloud.core.domain.user.CreateUser;
+import com.eager.questioncloud.core.domain.user.dto.CreateUser;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import jakarta.validation.constraints.NotBlank;
