+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/register/RegisterUserService.java
@@ -3,14 +3,14 @@ package com.eager.questioncloud.application.api.user.register;
 import com.eager.questioncloud.application.mail.EmailSender;
 import com.eager.questioncloud.core.domain.social.SocialAPIManager;
 import com.eager.questioncloud.core.domain.social.SocialPlatform;
-import com.eager.questioncloud.core.domain.user.AccountType;
-import com.eager.questioncloud.core.domain.user.CreateUser;
-import com.eager.questioncloud.core.domain.user.User;
-import com.eager.questioncloud.core.domain.user.UserAccountInformation;
-import com.eager.questioncloud.core.domain.user.UserInformation;
-import com.eager.questioncloud.core.domain.user.UserRepository;
-import com.eager.questioncloud.core.domain.user.UserStatus;
-import com.eager.questioncloud.core.domain.user.UserType;
+import com.eager.questioncloud.core.domain.user.dto.CreateUser;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.enums.UserStatus;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
+import com.eager.questioncloud.core.domain.user.model.User;
+import com.eager.questioncloud.core.domain.user.model.UserAccountInformation;
+import com.eager.questioncloud.core.domain.user.model.UserInformation;
 import com.eager.questioncloud.core.domain.verification.Email;
 import com.eager.questioncloud.core.domain.verification.EmailVerification;
 import com.eager.questioncloud.core.domain.verification.EmailVerificationProcessor;
