+++ qc-application/src/main/java/com/eager/questioncloud/application/security/JwtAuthenticationFilter.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.security;
 
 import com.eager.questioncloud.application.api.authentication.AuthenticationTokenManager;
-import com.eager.questioncloud.core.domain.user.UserRepository;
-import com.eager.questioncloud.core.domain.user.UserWithCreator;
+import com.eager.questioncloud.core.domain.user.dto.UserWithCreator;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
