+++ qc-application/src/main/java/com/eager/questioncloud/application/security/UserPrincipal.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.security;
 
 import com.eager.questioncloud.core.domain.creator.model.Creator;
-import com.eager.questioncloud.core.domain.user.User;
-import com.eager.questioncloud.core.domain.user.UserType;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.model.User;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
