+++ qc-application/src/test/java/com/eager/questioncloud/application/api/authentication/AuthenticationProcessorTest.java
@@ -5,14 +5,14 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 
 import com.eager.questioncloud.core.domain.social.SocialAPIManager;
-import com.eager.questioncloud.core.domain.user.AccountType;
-import com.eager.questioncloud.core.domain.user.CreateUser;
-import com.eager.questioncloud.core.domain.user.User;
-import com.eager.questioncloud.core.domain.user.UserAccountInformation;
-import com.eager.questioncloud.core.domain.user.UserInformation;
-import com.eager.questioncloud.core.domain.user.UserRepository;
-import com.eager.questioncloud.core.domain.user.UserStatus;
-import com.eager.questioncloud.core.domain.user.UserType;
+import com.eager.questioncloud.core.domain.user.dto.CreateUser;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.enums.UserStatus;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
+import com.eager.questioncloud.core.domain.user.model.User;
+import com.eager.questioncloud.core.domain.user.model.UserAccountInformation;
+import com.eager.questioncloud.core.domain.user.model.UserInformation;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import org.junit.jupiter.api.DisplayName;
