+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/cart/infrastructure/CartItemItemRepositoryImpl.java
@@ -4,7 +4,7 @@ package com.eager.questioncloud.core.domain.cart.infrastructure;
 import static com.eager.questioncloud.core.domain.cart.infrastructure.QCartItemEntity.cartItemEntity;
 import static com.eager.questioncloud.core.domain.creator.infrastructure.QCreatorEntity.creatorEntity;
 import static com.eager.questioncloud.core.domain.question.infrastructure.QQuestionEntity.questionEntity;
-import static com.eager.questioncloud.core.domain.user.QUserEntity.userEntity;
+import static com.eager.questioncloud.core.domain.user.infrastructure.QUserEntity.userEntity;
 
 import com.eager.questioncloud.core.domain.cart.dto.CartItemDetail;
 import com.eager.questioncloud.core.domain.cart.model.CartItem;
