+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/post/infrastructure/PostCommentRepositoryImpl.java
@@ -4,7 +4,7 @@ import static com.eager.questioncloud.core.domain.creator.infrastructure.QCreato
 import static com.eager.questioncloud.core.domain.post.infrastructure.QPostCommentEntity.postCommentEntity;
 import static com.eager.questioncloud.core.domain.post.infrastructure.QPostEntity.postEntity;
 import static com.eager.questioncloud.core.domain.question.infrastructure.QQuestionEntity.questionEntity;
-import static com.eager.questioncloud.core.domain.user.QUserEntity.userEntity;
+import static com.eager.questioncloud.core.domain.user.infrastructure.QUserEntity.userEntity;
 
 import com.eager.questioncloud.core.common.PagingInformation;
 import com.eager.questioncloud.core.domain.post.dto.PostCommentDetail;
