+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/social/SocialPlatform.java
@@ -1,6 +1,6 @@
 package com.eager.questioncloud.core.domain.social;
 
-import com.eager.questioncloud.core.domain.user.AccountType;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
 import lombok.Getter;
 
 @Getter
