+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/user/dto/MyInformation.java
@@ -1,5 +1,7 @@
-package com.eager.questioncloud.core.domain.user;
+package com.eager.questioncloud.core.domain.user.dto;
 
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.model.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
