+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/user/dto/UserWithCreator.java
@@ -1,6 +1,7 @@
-package com.eager.questioncloud.core.domain.user;
+package com.eager.questioncloud.core.domain.user.dto;
 
 import com.eager.questioncloud.core.domain.creator.model.Creator;
+import com.eager.questioncloud.core.domain.user.model.User;
 
 public record UserWithCreator(User user, Creator creator) {
 }
