+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/user/implement/PasswordProcessor.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.user;
+package com.eager.questioncloud.core.domain.user.implement;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
