+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/user/model/User.java
@@ -1,5 +1,8 @@
-package com.eager.questioncloud.core.domain.user;
+package com.eager.questioncloud.core.domain.user.model;
 
+import com.eager.questioncloud.core.domain.user.enums.UserStatus;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.implement.PasswordProcessor;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import com.eager.questioncloud.exception.NotVerificationUserException;
