+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/userquestion/UserQuestionRepositoryImpl.java
@@ -2,7 +2,7 @@ package com.eager.questioncloud.core.domain.userquestion;
 
 import static com.eager.questioncloud.core.domain.creator.infrastructure.QCreatorEntity.creatorEntity;
 import static com.eager.questioncloud.core.domain.question.infrastructure.QQuestionEntity.questionEntity;
-import static com.eager.questioncloud.core.domain.user.QUserEntity.userEntity;
+import static com.eager.questioncloud.core.domain.user.infrastructure.QUserEntity.userEntity;
 import static com.eager.questioncloud.core.domain.userquestion.QUserQuestionEntity.userQuestionEntity;
 
 import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
