+++ qc-core/src/test/java/com/eager/questioncloud/core/domain/user/UserRepositoryTest.java
@@ -2,6 +2,14 @@ package com.eager.questioncloud.core.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.eager.questioncloud.core.domain.user.dto.CreateUser;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.enums.UserStatus;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.infrastructure.UserRepository;
+import com.eager.questioncloud.core.domain.user.model.User;
+import com.eager.questioncloud.core.domain.user.model.UserAccountInformation;
+import com.eager.questioncloud.core.domain.user.model.UserInformation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
