+++ qc-core/src/test/java/com/eager/questioncloud/core/domain/user/UserTest.java
@@ -3,6 +3,13 @@ package com.eager.questioncloud.core.domain.user;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.eager.questioncloud.core.domain.user.dto.CreateUser;
+import com.eager.questioncloud.core.domain.user.enums.AccountType;
+import com.eager.questioncloud.core.domain.user.enums.UserStatus;
+import com.eager.questioncloud.core.domain.user.enums.UserType;
+import com.eager.questioncloud.core.domain.user.model.User;
+import com.eager.questioncloud.core.domain.user.model.UserAccountInformation;
+import com.eager.questioncloud.core.domain.user.model.UserInformation;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import com.eager.questioncloud.exception.NotVerificationUserException;
